Solution ready !
So, I was struggling with indexing out the elements from my 2-D array of cluster out of the loop which has to be displayed onto my 1-D array of Clusters. What I have learnt is that., it is very import to know the sequence or the order of 1-D array of cluster "elements" and keeping in mind that then only you should be indexing out the elements of 2-D array and then building onto a cluster again using inbuilt function inside cluster Pallette. This solution does not uses BUNDLE/UNBUNDLE to manipulate data update onto a cluster terminal. Look at the code and you shall understand it.